cmdlib: drop `force` parameter to `import_ostree_commit`
This was originally needed for `cosa sign` to force the re-import from
the tarball to ensure that the detached metadata was present in the tmp
repo after signing (3b192cc7a64a).

But now with the switch to `ostree container`, we operate directly on
the tmp repo and re-export the signed commit, so this is no longer
needed.

This also fixes `cosa generate-hashlist` also setting `force`, which I
don't think it actually needed.