cmdlib: mount `src/config-git` into supermin instead of `src/config` if present
The real repo is in `src/config-git` in a subdir setup and `src/config`
is just a relative symlink to the subdir in question.

This should allow us to use subdirs in the CI of openshift/os for
different RHEL/CentOS versions.

Long-term, we should probably move away from subdirs into something more
structured that doesn't require symlinking everything into the subdirs.