cmdlib: bump supermin VM memory to 4G
Something has changed recently which causes us to hit the ENOMEM issue
more easily now:

https://github.com/openshift/os/issues/594#issuecomment-1163549882

Mid-term, we could rework the compose so that only the OCI archive is
pulled through 9p rather than a full `pull-local`. Long-term, the fix is
to stop using 9p.

But for now to unblock CI, let's just bump the VM memory to 4G which
should help.