cmdlib: mount supermin root disk by UUID
On s390x we've seen the supermin VM get the two disks attached to
it when running buildextend-qemu mixed up such that the blank 10G
disk to be populated is vda and superman fails to start because it
can't mount a blank disk. Let's mount via filesystem UUID here.

Fixes https://github.com/coreos/coreos-assembler/issues/2941