cmdlib: use `repo-packages` for `overrides/rpm`
In addition to a lockfile, we can be even stronger here by using
repo-packages to tell rpm-ostree that the overridden packages must come
from the generated overrides repo.

This also allows one to override RPMs which themselves are specified as
`repo-packages` in the manifests without in most cases having to edit
them.

Related: https://github.com/coreos/rpm-ostree/issues/3789