cmd-build: move `--write-composejson-to` switch to cmdlib
Now that the variable is declared publicly, we can just add it to the
list of rpm-ostree args as part of the compose helper.

This has no functional change in the `cosa build` path. In the
`cosa fetch` path, this will add the switch but rpm-ostree will ignore
it since we're not actually writing a commit.

(cherry picked from commit 4be28b64564fc8a347c8d891cda17df7d4dcb83c)