platform/azure: delete storage group on failure
If e.g. we fail to create the storage account, we need to make sure we
delete the resource group.

Leaking is not a big deal if you also regularly run `ore azure gc`, but
combined with #2956, it was causing us issues because the same resource
group name was being reused. If an initial run failed to delete the
resource group due to a network flake, subsequent runs could also fail
trying to e.g. create storage resources because they already existed.
Not deleting the resource group there meant that later jobs could keep
failing the same way.