mantle/kola: extend --ssh-on-test-failure to work when systemd unit failrues are detected
It's frustrating when systemd unit failures are detected but you can't
just get a shell on the machine to investigate. This patch plumbs
through the SSHOnTestFailure information into the RuntimeConfig for the
machine object so we can use that information within CheckMachine to
dropt into a shell if desired by the user.