cmdlib: use `repo-packages` for `overrides/rpm` only as necessary
This is follow-up to 8c8b7338a ("cmdlib: use `repo-packages` for
`overrides/rpm`").

That commit changed semantics of `overrides/rpm` such that all packages
found are forced to be installed rather than only "preferred".

This revert the semantics back to what it used to be while still
allowing `repo-packages` to be overridden.

We do this by simply only adding a `repo-packages` override for the
packages which are found in the flattened manifest.