Drop golangci-lint from image build
Installing it is flaky and regularly breaks Prow and/or CoreOS CI. I
think the main issue for Prow is that the install script uses the GitHub
API anonymously and we get rate limited.

Looking at the official docs, they recommend using the GitHub Action
instead so let's do that in a follow-up (and have it run over the whole
repo instead of just the gangplank code).