create_disk: use -p when creating grub2 directory
The use of `-p` was dropped in
9330b89f4368b2e9f24f859a8fd64bfe91c09eea to appease some ShellCheck
errors.  (Notably https://www.shellcheck.net/wiki/SC2174)

But dropping the use of `-p` causes a fatal error when the directory
already exists.

This reintroduces the use of the `-p` flag and splits out setting the
permissions on the directory as a separate operation.

Closes openshift/os#896