ci: Move ShellCheck to GH actions
This is similar to
https://github.com/coreos/coreos-assembler/pull/2977
https://github.com/coreos/coreos-assembler/pull/2978

Today we're shipping code linting tools for shell and python
as part of the coreos-assembler image, which makes little sense.

Move `ShellCheck` to a separate CI flow run via GH actions.  For local
development, this are now manual instead of run as part of
`make check`.

(I tried to move python too but `pylint` depends on half the world
 being installed)