kola: Add support for `injectContainer`
Closes: https://github.com/coreos/coreos-assembler/issues/2952

For quite a while we've been generating a "base image" container
as part of our builds, but it has basically been entirely untested
outside of its processing and use by the ostree stack.

This adds a base framework such that external tests can request
the injection of the container (`.ociarchive`).

Now, we can (and *definitely* should) also pursue "natively" testing
this container image outside of kola, which is very virtual-machine
focused.

However, that will also require a lot of pipeline changes; this
approach allows us to natively "bridge" the worlds of cosa/kola
to test the container image.