Stop building gangplank (but keep the code), drop `gangplank/vendor`
Following the plan in https://github.com/coreos/coreos-assembler/issues/2860#issuecomment-1184327074

This will help us move the Go code to the toplevel; we're going from
*three* vendor/ directories to two.  Then when we merge the mantle/
bits in we'll go down to the sanity of one vendor/ tree.

To emphasize: There's a lot of good ideas in gangplank, we're going
to be pulling *some* of this code for sure, so keep the code around
to make that explicitly easier.