src/cmd-remote-build-container: check registry up front
Let's check if the image with this tag is in the registry before we
do any builds. In the case where we are pushing to a tag that includes
the git hash of the repo that was used to build it then we can consider
the image to have already been built and skip building it again.