cosalib/builds: add get_latest_for_arch function; use it in cmd-build
Right now cmd-build will just pick up whatever the latest symlink points
to and try to get previous build information from that, but what if
the absolute latest build doesn't have a build for the architecture we
are targetting? Currently cmd-build will fail with:

```
/usr/lib/coreos-assembler/cmd-build: line 163: /srv/builds/36.20220720.20.5/aarch64/meta.json: No such file or directory
error: failed to execute cmd-build: exit status 1
```

What we really need is to get the latest build for the given
architecture and use that in cmd-build instead.