buildextend-metal: Add options for rhcos secex build pipeline
Allow customizing qemu arguments and add the option to execute
genprotimg outside of cosa.
Add the option to run a post script after disk creation.

Background: To build RHCOS Secure Execution Images, we need to use a special
hostkey, which we do not have direct access to. Therefore we need to execute
extra steps in this case.

Signed-off-by: Jan Schintag <jan.schintag@de.ibm.com>