build.sh: freeze coreos-installer to v0.12.0
The latest coreos-installer dropped support for the legacy alias
`iso extract pack minimal-iso`, so this broke the 4.10 pipeline.

The core issue is that we run miniso packing using the coreos-installer
from cosa. We should fix this and make it use supermin to run the
packing step with the coreos-installer in the compose (like for osmet).

But for now, let's work around this by just freezing coreos-installer on
the version we're actually shipping in 4.10.

Closes: https://github.com/openshift/os/issues/916