osmet-pack: pass through coreos-installer args as is
Parse `PRETTY_NAME` from `cmd-buildextend-live` instead of relying on
`osmet-pack` to do it.

Once we have that, there's no reason anymore to have script-level
arguments in `osmet-pack`. We can just have `cmd-buildextend-live` drive
the actual set of arguments to use.

It's peculiar that `cmd-buildextend-live` now must be aware that
`/dev/disk/by-id/virtio-coreos` is a thing, but we'll formalize this in
the next patch.