cmd-buildextend-live: pack miniso with composed coreos-installer
Currently, we're packing the miniso using the coreos-installer from the
cosa image. This is not ideal however, because it introduces skew
between the packing side and the unpacking side, which may be older.

Fix this by using the new `runvm-coreos-installer` to pack with the
composed coreos-installer.

For more information, see:
https://github.com/coreos/coreos-assembler/pull/3009