cmd/remote-session: specify --pull=always when creating container
Otherwise our remotes use their local copy of the remote container
image and it never gets updated from the first time it was pulled.

With --pull=always no work is done if the container latest version
of the container is already local (i.e. it uses the cached local
copy), but the container image is fetched if there is a newer version.
This is what we want.