Dockerfile: allow root group to add root certificates
Similarly to the familiar `/etc/passwd` OpenShift hack, loosen up
restrictions on `/etc/pki/ca-trust` so that the unprivileged UID can add
root certificates to the platform.

This allows cosa to be used as is to build in environments where
additional root certificates may be necessary to fetch build inputs.

Add a script which can be used instead of `update-ca-trust` (which
assumes root with `CAP_DAC_OVERRIDE`) to update the primary bundles we
really care about for our stacks.

(cherry picked from commit ad62159b95a63846bdb683429ede86fbc444009f)