mantle/platform/qemu: fix seting of LIBGUESTFS env vars
This setting of cmd.Env was overriding the LIBGUESTFS_BACKEND=direct
that was being set just a few lines before. So the command was getting
run with just LIBGUESTFS_HV being set, which happens to have no
effect unless LIBGUESTFS_BACKEND=direct is set.

Fix that by appending to cmd.Env on the second modification.