mantle/platform/conf: dedupe added SSH keys
It's an error to have duplicate SSH keys in an Ignition config. In the
`kola spawn` path, it's possible for the same key to be added multiple
times if it's both in the SSH agent key list and in the default SSH key
location list (`~/.ssh/`).

Deduplicate the added keys by first passing it through a set.