Protect s390x secex filesystems using dm-verity instead of LUKS
LUKS with `--integrity` provided us with integrity protection so that we
were sure a malicious sysadmin couldn't have tampered with the disk
image. The problem however is that then we have to think about
protecting the generic per-cosa build LUKS keys.

We don't need confidentiality here, just integrity. Switch to using
dm-verity, which is made for this. Add new partitions to store the
hashes. Run `cryptsetup format` to populate them. Put the root and boot
partition root hashes in the kernel command line. Delete code related to
the previous LUKS approach.

On first boot, we will delete the hash partitions, and reprovision the
rootfs and bootfs on top of LUKS.

Co-authored-by: Nikita Dubrovskii <nikita@linux.ibm.com>