kola: Also support inline YAML for metadata
The *third* salvo in my battle against the duplicative and verbose
required comment metadata we've grown in fedora-coreos-config.

A key objection I have to it is that it requires specifying
the option *twice* - meaning it's easy for the actual options and
the comments to get out of date, etc.

Add support for YAML for metadata, which is just nicer to write:

- We support comments inline
- The data can span multiple Lines

So if we have comments to add, we can now do it in a better way.