create_disk: Fix s390x regression from previous change
On every architecture except s390x we rely on BLS i.e.
`sysroot.bootloader=none`.

However apparently there's a bug here if we try to deploy
with the `bootloader=zipl` from the start, instead of doing
it later.  I think that's an ostree bug, but for now rework
the code so that it works the same as before on s390x.

Closes: https://github.com/coreos/coreos-assembler/issues/3067