Move the schema/ package to the toplevel
Today, we have *three* Go modules in this repository:

- /
- schema/
- mantle/

And as of recently, both the toplevel / and mantle/ vendor the schema/.
This causes "vendor amplification", where a vendored dependency
of schema/ gets *triplicated*.

Now that gangplank/ is out of the picture, it makes sense to move
the bits from schema/ into the toplevel module.

This really helps make operating on coreos-assembler stuff
feel like a "first class" operation.

Also, we no longer have two copies of the schema file!