build-extensions-container: use `finalize-artifact`
We hit on the problem that current RHCOS pipeline uses a separate
volume for `builds/`, and we can't use `rename()` across that.
One thing the `finalize-artifact` wrapper does today is use
`mv` which will use an efficient `rename()` if possible, otherwise
fall back to a copy.

(We also now more consistently make the artifact read-only)