builds: Fix arch to use standard coreos arch function
A perfect example of the enormous technical debt we've accrued
in the python/shell/Go duplication, not to mention separate
Go modules for mantle and gangplank.

The gangplank code reimplemented logic to get the RPM architecture,
and missed converting `arm64` from Go's arch to `aarch64`.

Use the standard API we have for this in stream-metadata-go,
the same as mantle does.

This will fix extensions builds on aarch64.