Remove `tools/`, install schematyper on demand
We've had a few build/CI flakes due to the `go install` invocations
happening as part of a build.  We previously removed `gomarkdoc`,
this completes things and stops building/installing `schematyper`
by default.

It's only need for developers to regenerate the schema; we commit
the generated code, so we don't need to build or ship it by default.