Add `pipeline.yaml`
This is motivated by
https://github.com/coreos/fedora-coreos-pipeline/issues/618

Today the coreos-assembler default for `cosa compress` is `gzip`,
and the RHCOS pipeline relies on that.  The FCOS pipeline overrides
it.

My opinion is that this is exactly the kind of thing that should
be configured in the source git - it should not live in "the pipeline".

By having a new `pipeline.yaml` in this repository, we get closer
to the goal of having the default invocations of `cosa $x` do the
same thing as production pipelines.

If accepted I plan to extend this with numerous other things; most
notably for example the set of desired target artifacts i.e.

```
artifacts:
  - aws
  - metal
  - digitalocean
  - ...
```

etc.  Then we can have `cosa buildextend-all`, etc.

This PR includes some ugly bridging of Go logic to Python.