Move some fedora-coreos-releng-automation scripts here
We have some loose scripts in the fedora-coreos-releng-automation repo
used by the FCOS pipeline. It's always been awkward however: some of the
scripts import cosa Python/shell APIs, and the pipeline has to clone the
repo just to use the standalone scripts.

Initially I had thought we'd eventually find a way to have a sort of
sidecar container to make that cleaner, but... that never happened.
Another idea was a layered cosa image, but experience since then has
shown deriving the cosa image has many downsides.

Purely for practical purposes, let's just have them in cosa directly.
It's not an ideal fit but OTOH we already have a lot of FCOS-specific
things in here. This will allow us to clean up the pipeline code and the
automation repo.

I've renamed the scripts from their original names so that they're no
longer suffixed by `.py` since that's an implementation detail, and
prefixed by `fcos` or `fedmsg` to make clearer their scopes.

I've also added a link to the original repo at the top in case one wants
git history for those scripts.