cmd/remote-session: support `cosa remote-session create --workdir`
This allows the user to specifiy what directory in the remote container
we want the working directory (COSA directory) to be. This is useful
when you have a particular directory you are in on your local host and
you want to use the same directory in the container to keep things
symmetrical.

The first use of this will be in the pipeline where we'll set --workdir
equal to the Jenkins ${env.WORKSPACE}.