cmdlib: cleanup tmp builddir if allocated
The `runvm` function was written to support `tmp_builddir` being unset,
in which case it takes care of allocating a tmpdir itself. However, it
does not take care of deleting the tmpdir if it allocated it. These
directories can pile up in one's `tmp/` dir.

Fix this by tracking if we allocated the directory, and delete it on the
way out if so.