runvm-coreos-installer: run from workdir
Mount the workdir into the chroot and set it as the working directory
when running coreos-installer. This ensures that paths under the workdir
remain valid.

One hack we have to do here is create the canonical OSTree var subdirs
which are pointed at by toplevel symlinks so that `mkdir` doesn't get
confused by the broken symlink if it's under one of them.