runvm-coreos-installer: be less spammy
Runing in bash verbose mode the whole bit inside supermin that sets up
the chroot is a bit overkill. It adds a lot of output given that we run
this script three times when building the live ISO.

Let's only enable it for the actual coreos-installer invocation because
it's still useful to see the final arguments that were passed.