cmd-push-container-manifest: add container tags pushed to `meta.json`
For FCOS, we'll eventually want to add the oscontainer image in the
stream metadata. We'll want it to be tag-based so that by default, users
always use the latest release.

Add a `tags` key to the image definition for this.

We include all the tags instead of just e.g. the first one, because I
think it's good practice to have `meta.json` reference all the external
resources that it pushed to. This is important for example for GC.