build-extensions-container: various cleanups
Add more context about how the script is run.

Use `shift` on the same line we read arguments for consistency with the
rest of the codebase.

Factor out `localhost/extensions-container` into a variable.

Instead of `cd`ing into the config dir, pass it as the context dir
directly to podman.