create-legacy-oscontainer: rename command to `buildextend-legacy-oscontainer`
This way, it's consistent with all the other artifacts we build in the
pipeline which makes it easier to add support for. Don't retain the old
name since it's not used anywhere else yet.

(cherry picked from commit 03dd8da933722902b9775c823af68afa5187f774)

jlebon: Had to add command to bash entrypoint instead of go.