cmd-buildextend-legacy-oscontainer: don't override TMPDIR
We want `buildah` to use the default temporary directory instead of one
that lives on the 9p mount. Do this by stopping to pass `workdir` (not
to be confused with the cosa workdir; this is the directory which
previously contained the temporary containers-storage and tmp dir).

This whole thing is really hard to follow. But the nice thing is that we
should be able to get rid of it soon.

(cherry picked from commit 1f05e1d05db1b09bde82d37e794ae3666f3fe724)