cmd-generate-hashlist: use `lsinitrd --unpack` to unpack initrd
Rather than a custom `skipcpio | gunzip | cpio` pipeline, we can just
use `lsinitrd --unpack` here. It's simpler, but it also fixes the
command on systems that no longer use gzip for initramfs compression. It
also provides better error-handling (e.g. note that this command wasn't
failing when FCOS switched to zstd). Finally, this also fixes the fact
that `cpio` was run with `-v` which generated a lot of output in
pipeline logs.

For keylime, I'm actually not sure if this makes sense. The list of
initramfs files are hashed and included in the same JSON object keyed by
file path. This means that an initrd file could clash with a real root
file. The schema supports multiple hashes per keys, but the code isn't
appending, just replacing. To be investigated whether keylime really
wants the digests of files inside initramfses.