mantle: overhaul aliyun auth
The Aliyun SDK has a native file format it can consume for
credentials [1]. This commit proposes we use that file format
here and drop the custom JSON parsing that we're doing for Aliyun.

This commit also drops the toplevel `--config` argument in
`cmd-ore-wrapper` because Aliyun was the only consumer of that argument.

[1] https://github.com/aliyun/alibaba-cloud-sdk-go/blob/710fea6cf27cf855f301462a2adcf3399169f63c/docs/2-Client-EN.md#2-credentials-file