cmd-buildextend-live: avoid no-op `renameat(2)` for FCOS
On FCOS, the EFI directory is already correctly named in the src config
repo. So there's no need to do any renames at all. The previous code
still worked because `renameat(2)` with the src and dest being the same
file is a no-op.

This is a bit too implicit though. Let's just add a check to be more
clear. This avoids us also unnecessarily iterating over
`kargs_json['files']`.

While here, add a print statement when this logic kicks in.