Set `IMAGE_TYPE` when calling `prepare_build` from buildextend commands
This is necessary so that the `tmp_builddir`s allocated are distinct per
artifact run. Otherwise, running the two commands in parallel will stomp
on each other as they try to use the same dir.

`buildextend-metal` already does this, so we're missing
`buildextend-extensions-container` and `buildextend-legacy-oscontainer`.

I believe this is the reason why `buildextend-extensions-container` has
been failing in the RHCOS pipeline.

I avoided a larger cleanup here around enforcing `IMAGE_TYPE` (probably
should be renamed to e.g. `ARTIFACT`) getting set in `prepare_build`.
That requires more work because we should also make sure to consistently
clean up the dir at the end. Right now, we're relying on the fact that
the same dir gets cleaned and reused each time. We'll also need to
backport this so I wanted to keep the changes minimal.