drop tools and gangplank
We've already dropped these in the latest COSA code. Let's drop
them here because we're seeing build failures because go is trying
to pull from different source code repos and it's failing. Rather
than wait for the intermittent problem to resolve itself OR otherwise
have to workaround this, let's just drop the code since it's not needed.