cmd-compress: support `compressor` field in `image.yaml`
FCOS compresses its artifacts with xz, RHCOS with gzip. This is
currently encoded in the pipeline, but it should be driven at a lower
level so that we don't have to care about this in the pipeline and also
so developers don't have to think about passing the right switch
locally.

Extend `image.yaml` to support a `compressor` field for this. This will
be used by FCOS to override the default `gzip`.

This will allow us to drop a hack in the CoreOS pipeline.