cosalib/build: only take artifact lock when actually building it
We have this lock to prevent multiple `buildextend-${platform}` from
running concurrently. But it gets taken when the `_Build` object is
initialized rather than when we actually need to build it.

This clashes with our use of `cosa buildextend-aws --upload`, where the
artifact is already built and we just want to upload it. We should be
able to support multiple concurrent uploads to different AWS partitions.

Fix this by moving locking to when we actually build the artifact so we
don't take the lock during pure upload operations.

Note this is independent of the meta-update lock.

(cherry picked from commit 951d2496fca64889a1b3a9527afdf41f3e365975)