cmdlib: handle dupes when merging lists
It's the expected merge semantic and matches e.g. what Ignition config
merging does. Not deduping could technically be useful for kargs, but we
don't really have a use case for that so I didn't bother adding a e.g.
`allow_dupes` parameter.

Prep for next commit where we'll use this in a context where dupes are
expected.