cmdlib: merge inner lists when merging dicts
The expected semantic when merging two e.g. `meta.json` files is that
list keys are merged together, not that one clobbers the other. We
interestingly don't have many list types in the schema, but `amis` is
one of them and hits this right now because we upload to regular AWS and
AWS GovCloud in parallel. (In retrospect, `amis` should have probably
been a `dict` too, with the region as the key name. So a future cleanup
would be to migrate the schema to that model.)