cmd-push-container-manifest: fix wrong digest in meta.json with v2s2
When pushing with `--v2s2`, the digests will not match those in the
locally-created manifest because the format will be different. So we end
up putting invalid digests in the `meta.json`.

Surprisingly, `podman manifest create` does not support the `--format`
switch, so we can't create the manifest from the start so we get the
right digests.

Fix this by immediately doing a `skopeo inspect` of the remote repo
after pushing in v2s2 mode.

Related: https://github.com/containers/podman/issues/16603