plume/release: add support for --bucket-prefix
In the CoreOS pipeline, the pipeline configuration file supports
providing the S3 bucket and prefix. We make use of this in FCOS to
centralize the `prod/streams/${STREAM}` layout a bit more.

Let's have `plume release` take the bucket and prefix so we can stop
hardcoding the layout in this function and make it obvious from the call
that the invocation doesn't do anything to the bucket that falls outside
that prefix.

(cherry picked from commit 2eb51f2dcb09c1bf809b4b1c3d6d0db8650baeb5)