plume/release: drop doS3() code
Early in FCOS development, we were still debating whether the S3 objects
should be public or not. There was some initial agreement on uploading
first as private and then make specific builds public during the release
process, but in the end we've just always published in public from the
start.

It's not likely we'll change this anytime soon unless a good reason
comes up. So let's just delete the code here that makes the build
objects public since they already are. It'll live on in git at least if
we ever want to restore it.

Related: https://github.com/coreos/fedora-coreos-tracker/issues/189