plume: deprecate `release`, add `make-amis-public` and `update-release-index`
Focus for FCOS release logic has shifted to the `release` Jenkins job. I
think we could've gone the other way and put more of the release logic
like pushing containers and promoting GCP images into plume. We might
end up there eventually, but right now the pipeline is heavily centered
around doing cosa/mantle calls for specific things, and so it's really
awkward to have this generic `plume release` call that acts like a
blackbox and does multiple things under the hood.

Let's split the `release` command into two:
- `make-amis-public`: to make all the AMIs in a specific release public
- `update-release-index`: to inject new releases into the release index

Then it's really clear what each command does. This will also allow us
to get rid of a hack we're carrying for not calling this command for
RHCOS (though I'd like to make use of `make-amis-public` for RHCOS to
match the flow of FCOS).