cmd-buildextend-extensions: fix import_ostree_commit arg
The first arg is the workdir, not the OSTree repo path. The previous
version still worked because `import_ostree_commit` treated it as a
blank `tmp/` and repopulated the repo, which means we were paying the
cost of re-extracting the oscontainer and recompressing all the files
during the pull-local into the archive repo.

(cherry picked from commit a361f252ba0ba313a705d9ae87f1e9c24742e724)