cmd-generate-hashlist: rename to buildextend-hashlist-experimental
That way we can call it using the same pattern we call other buildextend
commands. It's not a bona-fide artifact (we don't add anything to
`meta.json` yet) because it's still experimental. But doing it this way
will allow us to easily skip running this on pipeliens where we don't
need it.

Motivated by issues we're currently hitting in RHCOS with generating the
hashlist being excruciatingly slow. Current theory is that the disk is
rotational and the random reads that the command triggers by checking
out the full commit and hashing everything is extremely expensive on
hard drives. I'd like to skip hashlists in RHCOS for now until we can
either move parts of it to tmpfs or ideally move to a builder with
better storage.

But regardless, this will also help us clean up in the pipeline how this
command is called.