devshell: Don't output status if not on a tty
This is just better behavior for the case of e.g.
`cosa run -x "somecmd" > out.txt` to execute a command and
capture the output - we don't want to intermix status stuff in
there.

Further, this makes it easier to test the code when not on
a tty (in e.g. CI).