plume/release: consistently add default region for `--region` switch
In the final command we define, we set the default region to the empty
string. This takes effect regardless of the command that actually gets
called and so we lose the default value of `us-east-1` on earlier
commands.

I think requiring the caller to pass a region would be cleaner, but we
can't break the deprecated `plume release` command yet.

Fixes d039e8830 ("plume: deprecate `release`, add `make-amis-public` and
`update-release-index`").