tests: choose latest y-stream versioning for rhcos upgrade
The cincinatti graph is used to determine the latest OCP release. It
was assumed that only the latest node would have no upgrade path in
the graph but that is not the case. When upgrades are blocked, there
can be two nodes that no have upgrade path. We need to choose the
newest one in the upgrade test.

(cherry picked from commit 48f57ce00e1698c5c22279e8142d81afe9658ca0)

jlebon: trivial conflict resolution in import list