tests: choose latest y-stream versioning for rhcos upgrade
The cincinatti graph is used to determine the latest OCP release. It
was assumed that only the latest node would have no upgrade path in
the graph but that is not the case. When upgrades are blocked, there
can be two nodes that no have upgrade path. We need to choose the
newest one in the upgrade test.