buildextend-live: drop `cio_ignore` from live ISO kargs
This switch came with initial support for s390x. It causes all devices
to be ignored except for the console. I think the rationale there is
that it allows the OS to boot faster. However, this will not work for us
because:
1. The ISO depends on being able to find the CD-ROM device containing
   itself so that it can mount the rootfs from it.
2. Installing to a target device from the live ISO requires that device
   to be visible; I think it's possible to dynamically unignore devices,
   with the `cio_ignore` tool, but this then implies that users would
   have to provide a pre-installation script to do this, which isn't a
   great UX to force.
3. It causes virtio-serial ports to not show up, on which we rely
   heavily in `kola testiso`.

Once the ISO supports karg modification, customers who really want to
trim down the list of devices to bring up can reintroduce the karg. But
ISTM it doesn't make sense to default to it.